If the phone is idle, don't let the user get stuck on the in-call screen.
If the InCallScreen comes to the foreground but no call is active, that
can mean one of two things:

  (1) We had just tried to make an outgoing call, but failed (like if
      there was no service or we were in airplane mode.)

      In this case we display an error dialog.

  (2) We came to the foreground for some reason *other* than making a new
      call.  For example, if the user manually turned the screen off and
      then back on *after* a failed outgoing call.

      In this case we currently display the in-call screen in a
      useless blank state :-(

This change fixes case (2) to simply bail out, since there's no reason to
remain on the in-call screen if the phone is totally idle.  This change
has no effect on case (1).

(I originally developed+tested this in froyo, but decided to defer it to
gingerbread since we were pretty late in the froyo schedule, and this bug
only happens in really obscure scenarios anyway.)

Tested:
  (1) Scenarios from bug 2072046.  Error dialogs *are* visible when the
      error first occurs, but if you then turn the screen off and back on,
      the in-call UI dismisses itself.
  (2) Normal outgoing and incoming calls: no change.

Bug: 2072046
Change-Id: Id098c7ed3aec2c3c7f9586331ecf5e4f5200d0ee