Allow incoming calls in Emergency Callback Mode, even if not provisioned
Normally the phone app rejects any incoming calls if the device isn't
provisioned (i.e. before you've run thru the SetupWizard).

But since you *can* make an emergency call before completing the
SetupWizard, we also need to allow you to receive incoming calls in
emergency callback mode, regardless of whether the phone is provisioned.

(Emergency callback mode is a CDMA-only feature.  The device goes into this
mode for the next 5 minutes after you make an outgoing call to an emergency
number.)

This change was originally a patch from Reli.Tal@motorola.com in
b/2489602.  That was a patch against MotoBlur code, though, so I had to
apply it manually and throw out a bunch of stuff that didn't apply to our
tree.

Tested:
  - Passion:
    - incoming call on provisioned device: worked fine (the device rang)
    - incoming call on freshly-wiped device before SetupWizard: rejected

  - Sholes:
    - incoming call on provisioned device: worked fine
    - incoming call on freshly-wiped device before SetupWizard: rejected
    - incoming call on freshly-wiped device while in ECM: worked fine

      (I did that last test by using the Extended NAM menu to make "666" be
      an emergency number.  I then wiped the device, pressed "Emergency
      dial" from the first SetupWizard screen, dialed 666, and then hung up
      to get into ECM.  Then I called the device, and the incoming call
      *did* ring.)

Bug: 2489602
Change-Id: Ibb1fc40be73de9ac638c02c54930ec91d6e05cb0