Use ViewStubs for DTMF Dialer
Fix Bug: 2139530
There are two dtmf dialers, one that can appear directly
on the in call screen for devices with a proximity sensor,
and another that appears in a drawer for devices without
a proximity sensor.  Prior to this change, both were
automatically inflated when the InCallScreen is created,
which is a waste of memory and time.
Now, both are inside view stubs, and only one is inflated.

Change-Id: I7522e03dd1e856066164b7b2f5efffa8c138e1d5