Update patch set 1
Patch Set 1: No score; Looks good to me, but someone else must approve

This is a pretty minimal and safe change, and it looks reasonable to me.

It does depend on the lower-level changes (i.e. in telephony and in the
ril) that add the new value Connection.DisconnectCause.UNOBTAINABLE_NUMBER
and propagate it up to the application, though.  As long as the telephony
guys are OK with *those* changes, then this UI change looks good to me.

Patch-set: 1
Reviewer: David Brown <1000358@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1
Label: Verified=0