Update patch set 1
Patch Set 1: No score; Looks good to me, approved

In response to David Marques' question, AFAIK the behavior of GPRS/3G data when FDN is enabled is not specified.

Several years ago, at a previous job, I worked with some EGPRS radio modules that would not allow data access with FDN enabled, unless there was an FDN entry of "*99#" on the SIM. Since we were building a data-enabled phone, we had to write code to add a "*99#" entry to the SIM FDN if not present, and then hide that entry from the user. Big hassle.

I just verified that the Nexus One *will* allow data access with FDN enabled. IMHO this is the correct behavior for a smartphone. I hope there aren't any Android devices out there which require a "*99#" entry in the FDN list in order to create a data connection with FDN enabled, but it is a possibility to be aware of.

Regards,
Jake

Patch-set: 1
Reviewer: Jake Hamby <1003544@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Verified=0