Cleanup CDMA versus GSM behaviors in DTMF dialer
Add a new class called TelephonyCapabilities which is where we
can consolidate code that says "if (CDMA)" or "if (GSM)".
This is intended to answer questions like "does this
device support OTA provisioning?", so that the CDMA and
GSM knowlege does not have to be spinkled throughout
the Phone App.  Ultimately, these questions should
be actually answered by the telephone layer.

Remove differences in behavior between GSM and CDMA phones
in the DTMFTwelveKeyDialer.  There were many cases where the
code branched depending on whether the device was CDMA or
GSM, but then ultimately did the same thing.  The only
actual behavior difference between CDMA and GSM was whether
to use short or long DTMF tones, this policy question is
now answered by the TelephonyCapabilties.

Also, get rid of unneccessary behavioral differences
in the InCallScreen handlePostOnDialChars.  On CDMA devices,
the pause dialog gives the option to cancel, but this is
not available on GSM devices.  This functionality should
not have anything to do with GSM versus CDMA, so now this option
is available regardless of phone type.

Change-Id: Ie3a7e2974b06674935079c4defcbabe6d2d4864c