get rid of duplicated code in placeCall() and placeCallVia()
These two functions had mostly the same functionality, but
structured differently.  Now these functions have been
merged into placeCallInternal().
This also fixes Bug: 2667919, which was that the speaker
phone was not activated when in the car dock and
using google voice.  This was caused by a line
missing from placeCallVia().

Tested on Passion:
regular numbers, with and without google voice
regular numbers in car dock, with and without google voice
phone numbers with pause and wait in the post dial numbers
MMI codes
USSD codes

Tested on Sholes:
regular numbers, with and without google voice
phone numbers with pause and wait in the post dial numbers
emergency numbers

Change-Id: I5630f09cea8f27357a72b007aa7baf0277d59e96