Make the ongoing call notification selectable
Fix Bug: 2572519
When the ongoing call notification is selected,
most of the notification layout did not change color and
did not appear to actually be selected.
Unlike other notifications, the ongoing call notification
uses a custom layout because it needs to display the
call duration.  There were several differences between
the ongoing call layout and the standard layout
in status_bar_latest_event_content.xml, in particular
that the ongoing call notification specified
a background color.  It is now essentially the same as
the standard layout, except for the differences neccessary
to display the call duration.

Change-Id: I69772ca59ea8e8c3c89de22a2459b3469e575751