CDMA/GSM cleanup in CallNotifier
Previously, the CallNotifier would register for certain
messages if the phone of was a phoneType that
was expected to provide that type of message.
For example, we would only register for
displayInfo and signalInfo messages on CDMA
devices.
Now, CallNotifier will always register for these
messages, regardless of phone type.  If the
current phone does not actually support that
functionality, then it will simply never deliver
that message.  This simplifies the code in the
phone app.

Change-Id: I77a7b0895010b1f402f814fbf16e13d9157a01fa