Clean up "if (CDMA)" checks for logging emergency numbers.
On some devices we disable logging of emergency numbers in the call log.
The code previously said "if (CDMA)", but this behavior is actually
Sholes-specific since it's really a VZW requirement.

Now, this behavior is controlled by a platform-specific resource called
"allow_emergency_numbers_in_call_log".  The default is true; I'll set it
to false in the vendor/moto/sholes resource overlay in a separate change.

Also use supportsOtasp() instead of "if (CDMA)" in one more place.

Change-Id: I7608102a32f489575f8f8c6f6ef4d15daf930628