Fix a Sholes-specific "stuck on the in-call screen while idle" bug
Fix one last loose end of bug 2072046 that I missed in my earlier fix
<https://android-git.corp.google.com/g/49023>.

The inital fix worked fine for GSM devices, but you could still get stuck
on the in-call screen on a Sholes device because of a long-standing bug
that caused syncWithPhoneState() to ALWAYS return "SUCCESS" on CDMA
phones, even if the phone truly was idle.

This change fixes that bug.  We now consider the phone to be "in use" only
if there's an active call, or if an MMI is in progress.

Tested: on Sholes, ran the "ultra-simple scenario" from bug 2072046,
verified that after unlocking you get bounced back to contacts (or
wherever you started from) after step (5).

Bug: 2072046
Change-Id: I7bc02b8d414faa264644ed67bc80d4269f32f681