Clean up one last reference to KeyguardLock APIs in the phone app.
Long ago (like in donut and earlier), the phone app needed to use the
KeyguardLock.disableKeyguard() and reenableKeyguard() calls to allow the
"SIM network unlock" dialog to be visible even on a locked device.
Those APIs are now deprecated, and in fact will no longer work in certain
states (see bug 2737842.)

However the SIM network unlock dialog doesn't actually *need* to be
displayed on top of the lock screen in the first place.  It's a dialog
that you only get once, when the device first boots, and it's perfectly OK
if the user needs to enter his lock pattern (or dismiss the keyguard)
first, before we show the dialog.

So this change simply removes all KeyguardLock calls from the IccPanel
class, just to make it clear that the phone app no longer uses those APIs.

I also added a TODO comment about why it's hard to preserve the original
behavior: We can't just use the FLAG_DISMISS_KEYGUARD and/or
FLAG_SHOW_WHEN_LOCKED window flags, since since we're a Dialog (not an
Activity), and the framework won't ever let a dialog hide the keyguard
(because there could possibly be stuff behind it that shouldn't be seen.)

Tested: on Passion, using a temporary hack to make the wired headset
  button trigger a SIM_NETWORK_LOCKED event.  If the device is unlocked,
  the "SIM network unlock" appears immediately.  And if the device is
  locked, or if the keyguard is visible, the dialog appears as soon as you
  enter your pattern or dismiss the keyguard.

Bug: 2737842

Change-Id: Iace0eba81437ba40166c773e6cd377726d923afe