Update the phone app's status bar notification even while ringing
The phone app now puts up a notification any time the phone is OFFHOOK
*or* RINGING, rather than only in the OFFHOOK state.

This is a first step toward a new notification-based incoming call UI,
which we're planning to use if a call comes in while a non-interruptable
("immersive") activity is in the foreground.

Side note: this behavior is also useful in the (very obscure) case where
you somehow manage to escape the incoming-call UI while a call is still
ringing; see bug 2248148 for an example.  If this happens, you can now
select the "incoming call" notification to return to the incoming-call UI.

Also removed some old dead code that was trying to update the "expanded
view" RemoteViews *after* sending it over to the NotificationMgr (which
has no effect.)  Now, when the caller-id query finishes, we just
completely regenerate the in-call notification.

Tested: on Passion, using a temporary hack that let me press BACK to
escape the incoming-call UI.  After doing that, verified that the new
"incoming call" notification looked correct, and (when selected) took me
back to the incoming-call UI.

Bug: 2768760

Change-Id: I57bb072c746e944608324e8ed6cd1b230fdb96ce