Don't post the "incoming call" notification till we're ready.
Remove the call to NotificationMgr.updateInCallNotification() that was
happening directly from CallNotifier.onNewRingingConnection().

At that point we don't want to post a status bar notification (yet), since
we still need to run the caller-id query (and we may ultimately not even
ring at all, if the "send directly to voicemail" flag is set.)

(This is also why incoming calls would show up in the mini-me with a raw
phone number like "16505551212" rather than a contact name.)

We now update the notification -- and potentially launch the InCallScreen --
only from the showIncomingCall() method, which runs when the caller-id
query completes or times out.

Also add a bit more logging for the overall updateInCallNotification()
sequence.  (Like before, only for userdebug or eng builds.)

Tested: incoming calls, both immersive and non-immersive.  Also confirmed
        the contact name (not raw phone number) now shows up in the mini-me.

Bug: 2768760
Change-Id: Iccbc88f224377d967945c260dc1416b5e44faaa0