Don't try to go to the call log on devices with no call log
Clean up some code in the phone app that assumed the existence of the
"Call Log" screen, and would lead to bad behavior on non-voice-capable
devices where there's no such thing as the Call log.

(1) After a call ends, in situations where we're supposed to go to the call
    log, don't crash if there's nobody to handle the call log intent.

(2) When there's a missed call, don't bother posting the missed call
    notification at all if there's no call log.

and also:

(3) Sanity check the other startActivity() calls in the phone app to confirm
    that there are no other lurking problems (like bug 3056334, or any other
    cases where future products might not have all same the built-in apps as
    current ones.)

    (Result: Fixed two internal helper methods where we were manually
    bringing up the InCallScreen.)

NOTE that on Stingray you shouldn't have been able to get to the phone app
in the first place(!)

But there are currently some ways this can happen (see bug 3038102 and bug
3116625 for a couple of examples), so this fix really just makes the code
more robust in case you *do* somehow get to the phone UI on a
non-voice-capable device.

Bug: 3056334

Change-Id: I8ab41398a3a096a4462bbc5fc08aca56eae49376