Update patch set 5
Patch Set 5:

Jaikumar
We can share both the APK and the source code of the test application. Please note that this is a 'unit-test' app to demonstrate the Virtual-call feature. 

Let me know how you want me to send this code as this is not part of any GIT project.

Regards
/Kausik

Patch-set: 5