Update patch set 5
Patch Set 5:

I got the test case thanks. This is good for now, but we may have to change before exposing the API.

I am currently collapsing the AudioState change Intent with the HeadsetState change Intent to make it similar to A2DP.
So that we can expose a common interface.

When you are on a virtual call and you play music does this work ?

The reason I ask is AudioManager should know whether SCO is connected or out. we didn't make audio manager changes to handle the Virtual Call Audio state intent.

AudioManager code has these modes like Mode_IN_CALL etc.

I believe we will need to collapse Virtual Audio State intent with the Headset state change and have audio manager expose APIs to know whether we are in a virtual call or normal call. Like in a voice recognition case, the sco channel is opened before the call is connected. The call might not be connected and the user might just close the dialer.

Patch-set: 5