Move PhoneUtils.setAudioMode().
As different audio modes need to be applied for different phone types, (for example,
SIP call needs to be in MODE_NORMAL while GSM call in MODE_IN_CALL,) we move the
logic to CallManager. A side benefit is that since CallManager knows the current
phone state, PhoneUtils does not need to keep the audio control state (sAudioBehaviourState)
anymore.

The companion CL is at https://android-git/g/#change,61553

Change-Id: Ibb40e1853f2e5818b4b28c976f5756f567b3de41