Go to the Call log after *all* calls, even incoming calls.
As requested in Bug 2844990, as long as the screen isn't locked, we
now *always* go to the call log when a call ends, regardless of
whether it was originally incoming or outgoing.

(Previously, after an outgoing call we'd take you the Call log, and
after an incoming call we'd just return to whatever you were doing
when the call came in.)

Note we *don't* go to the Call log if the device is locked, since you
would just see the lockscreen instead of the call log anyway.  (And in
particular, if the device has a lock pattern set and a call comes in
while the screen is off, we *won't* go to the call log afterwards
since the device will be locked at that point.)

Bug: 2844990
Change-Id: I8211bea8f94b3bd4da7211cc42629dee55f57423