Update patch set 1
Patch Set 1:

Is there a carrier/compliance/regulatory requirement to honor this bit?  If not I'd rather leave the menu intact - gives more control to the user.  (And if it is, I guess we've been failing that requirement for while now?)

If it's a req only for some carriers, I would suggest using a config var in conjunction with the bit in EF_CSP, so that the EF_CSP is ignored unless explicitly required for a particular build.

Patch-set: 1
Label: Verified=0