Don't destroy+recreate the InCallScreen on docking or undocking.
Add "uiMode" to android:configChanges for the InCallScreen, which will
prevent the InCallScreen from being destroyed and recreated when you dock
or un-dock the device while in-call.  (The InCallScreen is designed to
*never* be destroyed once it's initially created.)

Also add a skeleton onConfigurationChanged() method for future use (like
when we eventually allow landscape mode for the in-call UI -- see bug 2089513.)
It has no effect right now, since *currently* the InCallScreen looks
exactly the same regardless of configuration.

Side note: This fix is peripherally related to
Bug: 3028132
which was an out-of-memory crash in the phone process.  The crash happened
in InCallScreen.onCreate(), and in the logs I noticed that the
InCallScreen was being re-created because of a dock event (which isn't
supposed to happen.)

I don't think this is the root cause of bug 3028132, though, since the
logs on that bug show only a single InCallScreen destroy/create sequence,
which does *not* do enough allocation to run the process out of memory all
by itself.

TESTED: on passion, docked and undocked repeately during a call, confirmed
the activity was not destroyed and recreated, and that the
onConfigurationChanged() method was correctly called.

Change-Id: I26039588878b172607ac4da7130533f0b77f8e93