SIM card customizations
Customization options are increased by introducing a way to read specific operator
defined flags from the SIM card. Operators can store information on the SIM card that
can be read at run-time or at phone start up. This enables the possibility to use different
settings depending on the information stored on the SIM card. An operator can for instance
have different APN settings depending on whether the SIM card is pre-pay or post-pay.

Change-Id: I0dffd2616968070c65367c7fb0b03f2b93a28c4e