Suppress "sip:" prefix for SIP addresses
When displaying a SIP address in the CallCard, strip off the "sip:" prefix
if present.

(Note: we don't yet store the "sip:" prefix in the contacts database, but
we might start doing that soon.  This change makes us ready for that.)

Also fix a potential NPE in some SIP-specific InCallScreen code (which
actually showed up (on a different kind of device!) in bug 3075925.)

Change-Id: Ie0ca355e0221463b88f1921aa6d6c3a64b19d0a3