Ignored or rejected calls should return to the previous activity, not the call log
In Froyo MR2, as part of the fix for bug 2844990, we updated the in-call
UI to *always* go to the call log when a call disconnects, regardless of
whether it was an incoming or outgoing call.  Unfortunately this meant we
would go to the call log even for rejected or missed incoming calls, which
we *didn't* indend to do.

So this change basically brings back the code I removed in change
https://android-git.corp.google.com/g/62821 , but with correct logic this
time: We now look at the DisconnectCause of the call that just ended, and
we explicitly *don't* go to the call log after INCOMING_MISSED or
INCOMING_REJECTED calls.

TESTED:

  - outgoing calls:  go to the call log when done

  - incoming call, answered then locally hung up: go to the call log when done
  - incoming call, answered then other end hung up: go to the call log when done

  - incoming call, rejected: return to previous activity when done
  - incoming call, ignored: return to previous activity when done

Bug: 3077420
Change-Id: Ie223a1e96d43f82fbd1610c1ce6074c5215bd787