CallFeaturesSetting: Fix one last showDialog() crash
We crashed because we tried to show a dialog (in response to an
asynchronous event) but our activity had already been torn down.

This exact issue has come up before (see bug 2412130).  The fix was to add
a new method called showDialogIfForeground(), which is basically a safe
version of showDialog() that will silently do nothing if we're not in the
foreground.

Unfortunately, after that fix went in, someone else added code that still
used the *unsafe* version of showDialog(), which ended up causing this bug
again :-(

We now use showDialogIfForeground() everywhere.

Bug: 3103109
Change-Id: Iba7fa55b49e4d1132a8ba6978027dde0f20edb9a