Enable logging of InCallScreen creation/deletion in all builds
In normal operation, the InCallScreen should only ever get created once,
and should *never* get destroyed, since we never allow it to finish().

So increase the logging verbosity in onCreate and onDestroy to make it
clear (even in user builds) if we're getting created/destroyed more than
expected.  This is helpful for debugging OOM issues like bug 3100957.

Bug: 3100957
Change-Id: I5dbfe14bcdeb89d0d4f32512638f7da4371f6283