Mark test strings as non-translatable
The apps/Phone/tests package is test code, so nothing there ever needs to
be translated.  So rename strings.xml to donottranslate_strings.xml so
that the translation process will skip it.

(Right now that file only has one string in it, but I'm about to add a few
more, so let's do this cleanup first...)

Also:
  - Actually start using that one string
  - Delete the bogus values-*/strings.xml here which should never have
    been created in the first place

Change-Id: I89b7bd09cd2f6eb1bd95abc3bd1890ea3362becb