Add CALL/DIAL test activity.
This is a test activity that mimics the behavior of 3rd party apps firing
off CALL and DIAL intents, which will be useful when working on bugs
3116684 and/or 3116625 (which cover handling DIAL and CALL gracefully on
non-voice-capable devices.)

Note this activity is part of com.android.phone.tests, which is not
installed by default, and it's not in the LAUNCHER category either.  So to
run it you need to manually build apps/Phone and do an "adb sync", and
then "adb shell am start -n com.android.phone.tests/.CallDialTest".

Bug: 3116625

Change-Id: Ibadbd36bdb06121611a501acfa20ad42ba3f38d7