Force incoming-call UI to come up for call-waiting calls.
Pre-gingerbread, when an incoming call came in we'd manually launch the
incoming call UI via a startActivity() call.  But now, we launch it by
posting a Notification, using the "fullScreenIntent" field to specify the
activity to launch.

But there's a quirk in the NotificationManager's behavior: it will
*ignore* the fullScreenIntent field if you post a new Notification that
matches the ID of one that already exists.  Unfortunately this is exactly
what happens when you get an incoming call-waiting call, since an "ongoing
call" notification is already visible.  So the NotificationManager never
launches our fullScreenIntent, which means that the incoming call UI
doesn't come up :-(

(In most cases this isn't a major problem: the InCallScreen is usually the
foreground activity while you're in a call, so you'll still see the
incoming call UI when you pull the phone away from your face after hearing
the call-waiting tone.  But if you bail out of the in-call UI by pressing
HOME during a call, you're hosed: the incoming call UI will *not* come up
automatically.  The only way to answer the call is to manually go to the
InCallScreen via the notification bar, which is pretty much
undiscoverable.)

This change works around the NotificationManager behavior by explicitly
canceling and then re-posting the in-call notification, only in this
specific case.

TESTED:
  - Regular incoming call while screen on
  - Regular incoming call while screen off
  - Incoming call-waiting call while screen on, still on InCallScreen
  - Incoming call-waiting call while screen on, after bailing out to Home
  - Incoming call-waiting call while screen off, still on InCallScreen
  - Incoming call-waiting call while screen off, after bailing out to Home
  - Incoming call-waiting call while screen off due to prox sensor
The incoming-call UI correctly appeared onscreen in all cases.

Bug: 3184149
Change-Id: Ic8469fdbe7733134f3f52a439c8b4b5658ac0322