DO NOT MERGE: Check for incoming connections state before making an outgoing connection.
Sometimes SDP record is not available, so we delay connection.
In the meantime an incoming connection can come and we get connected.
We then try to connect, no sdp records, we just set the state to disconnected.
Technically we are connected :-)

This code desperately needs a state machine.

Fixes bug: 3281016

Change-Id: Ie1a7ad53595f225cfe38eb71ba29759ebc01544c