Update patch set 6
Patch Set 6:

Kaushik: Since its a public API we need through another team which reviews public APIs. I am sure we went through this before but I can't find it in the review

One concern was we are exposing the intent. VIRTUAL_CALL_STATE_CHANGE. 

The only use case for the app to know whether its a virtual call or not, right ? 

Is there any other use case for this ? 

Another better way of doing this would be to have an extra in the AUDIO_STATE_CHANGE intent which tells what was the reason for the audio state change - cellular call or voice recognition or virtual voice call.

What was the exact use case for adding this extra intent ?

Patch-set: 6