Update patch set 6
Patch Set 6:

Jaikumar
Yes, we did discuss this during the review earlier. We could merge it into AUDIO_STATE_CHANGE intent, but I chose not to for the following reasons:	

a.) AUDIO_STATE_CHANGE intent may already be used by several applications, so overloading this with Virtual Call and sending the same intents especially when switching between cellular-call and virtual-call may create confusion. Keeping this separate allows for backward compatibility with existing applications.

b.) CALL_TRANSFERRED state may be relevant for VirtualCall but not for cellular call/voice recognition. Applications that need VirtualCall API could either play the audio locally or terminate the audio on the CALL_TRANSFERRED. Rather than forcing this state into AUDIO_STATE_CHANGE kept a separate VIRTUALCALL_STATE_CHANGE.
 
Primary reason for keeping it separate is (a). So if it is OK to expect Application Developers to update to the new APIs, then we could merge the VIRTUAL_CALL_STATE_CHANGE intent into AUDIO_STATE_CHANGE intent. I hope this clarifies. 

/Kausik

Patch-set: 6