Update patch set 6
Patch Set 6:

I understand your point. But I am trying to reduce the number of intents we expose here. If we go down this route, if some app in the future needs to know whether its a cellular call (for which we don't have any intent) or a voice recognition or a virtual call - we would need 3 intents.

We can add an extra to the AUDIO_STATE_CHANGE intent. Its not a public API. And only those apps which care about this extra need to be updated.

CALL TRANSFERRED is only for the case when the SCO gets disconnected right ? So why can't we just use AUDIO_STATE_DISCONNECTED ? We use AUDIO_STATE_DISCONNECTED in both cases for a cellular call - Audio transferred and call ended from remote case

Patch-set: 6