Ignore incoming calls (at the UI layer) on non-voice-capable devices
This change updates the phone app to reject incoming calls, and ignore MWI
events, on non-voice-capable devices.

This is not the "real fix" for bug 3038102, though.  This change does
prevent the symptoms (i.e. ringtone playing and voicemail notifications
showing up on Stingray), but the real fix is that incoming calls and MWI
events should be blocked at the telephony layer in the first place.

Bug: 3038102
Change-Id: Ife70782e15406c5e2b3a3286042a34adf03672dc