Use CallManager.getMute() in PhoneUtils.
Currently, PhoneUtils.getMute() only get mute state from the foreground phone.
But if the fg phone is idle while a bg phone is off-hook, then we should get
the mute state from the bg phone. All the logic should be done in
CallManager.getMute() so PhoneUtils just needs to call that.

Bug: 3323789
Change-Id: I21ad6df19aa82f720e746f8d3f6a4ba5933c639e