Added fix for race condition in ringer notification on CDMA variant
In CDMA during a incoming call, RIL sends out a
UNSOL_RESPONSE_CALL_STATE_CHANGED message which triggers a
GET_CURRENT_CALL. Since we are already getting the current call.
RIL sends out UNSOL_CALL_RING right after.

This causes the CallNotifier to think that it is a repeat ring,
so it calls ringer.ring() before the custom ringtone query is
finished.

The RIL sequence is different in GSM, it doesn't send out a
CALL_STATE_CHANGE first. So this problem doesn't happen in GSM.
So set a variable to wait for the custom ringtone query to finish
or time out before processing a Repeat ring request in CallNotifier.

Change-Id: I245cffc66583138a33212d63f19db8ed027b1e1e