Make Phone App's android:label more meaningful
The Phone App's android:label is currently the string "Dialer", which is a
relic left over from the days when the dialer UI used to be part of the
phone app.

But today, you only ever see that label in settings UIs like "Manage
Applications" and "Battery Use", where the string "Dialer" makes no sense,
especially on non-voice-capable devices.

So we now call it "Phone" on phones and "Mobile Data" on tablets, to
reflect more accurately what the phone app actually does.

Also cleaned up a couple of unused strings.

Bug: 3378543
Change-Id: Ib77b2df74e1537a6eb2eab8c2f63e87f2a7b5fa8