Fix issue 3515250: end chat while starting BT SCO
There is a potential deadlock if the outgoing SCO socket
thread (BluetoothHandsfree.ScoSocketConnectThread) is interrupted
while the socket is being connected. If the socket is successfully
connected, the thread tries to update the audio state by calling
setAudioState() which is a synchronized method. At the same time,
the thread that requested the interruption is executing audioOff()
which is also a synchronized method and is waiting on join() on
the SCO socket thread => deadlock.

The fix consists in aborting SCO audio setup in the SCO socket thread
if an interrupt is detected when returning from connect().

Also fixed a similar deadlock for incoming SCO socket.

Change-Id: Idde547dd113e58d105338003d9e8ebb033ca9dd6