Fix issue 3515250 again: end chat and BT SCO start
The fix in commit e28e6bd3 was not complete and some deadlock
possiblities were left.

This fix consists in simplifying the locking mechanism by removing
mScoLock and only synchonize on the BluetoothHandsfree object.

Also do not wait for the SCO connection thread to exit when
shuting down in audioOff(): as this method can be called from the UI
thread and the SCO connection can take several seconds we should not
block on join().

Change-Id: Ie11dd0031661397cfa385baaf34b9344d1e69aaf