InCallScreen: call finish() instead of moveTaskToBack() after a failed call
Background: To improve performance when launching the in-call UI, we never
actually call finish() on it.  Instead, when a call ends we simply move
the InCallScreen to the back of the activity stack.  (This prevents the
ActivityManager from destroying it, meaning that we won't need to re-run
onCreate() the next time we get launched.)

But this optimization has a bad side effect: when launching the
InCallScreen multiple times in quick succession (like when there's an
error placing an outgoing call followed by an automatic retry), the
InCallScreen sometimes gets left in an inconsistent state and the *2nd*
launch attempt fails.

(This sequence is quite rare, but it can happen in one very critical
scenario: making an emergency call while in airplane mode, where the first
call fails because the radio is off, and then we retry after turning the
radio on.  In that scenario it's very bad for the 2nd call attempt to
fail.)

Here's the workaround: in the (uncommon) case where we're exiting the
InCallScreen screen after a failed call attempt, we explicitly call
Activity.finish() instead of moveTaskToBack().  That means we *will* have
to re-create the InCallScreen from scratch for the next call, which adds a
couple hundred msec to the startup time, but at least we can be sure the
activity will launch properly.

This change has no effect on the behavior of the InCallScreen for regular
incoming or outgoing calls.

Bug: 4172599
Change-Id: I9499709ed89b519254386e98ffe3b09461a605f6