replace Settings activation item with data service item on LTE devices
LTE devices don't require activation. Instead, we'll give the user a way
to launch the browser on the account management page.

The URL of the page to launch is obtained from Gservices (via Secure Settings.)
The name of the Settings item is a generic "Set up data services" here, but
will be overlaid for the Moto/VZW LTE tablet to be "Verizon data services"

bug: 4100048
Change-Id: I2440c0b3563eed2b9ff2e21be6319a6b55b5bc0e