Introduce CallController (Step 1 of InCallScreen redesign)
This is the first step toward pulling "call control" out of the
InCallScreen, and leaving the InCallScreen responsible for only the
onscreen in-call UI.

Design doc:  http://wiki/Main/InCallScreenRedesign

The CallController is a singleton object which acts as the interface to
the telephony layer for all user-initiated telephony functionality, like
making outgoing calls.

In particular, any functionality which needs to happen whether or not
the InCallScreen is currently visible belongs in the CallController,
*not* the InCallScreen.

This change introduces the CallController class, and moves the entire
outgoing call sequence (basically placeCall() and its helper methods)
from the InCallScreen to the CallController.

Also, this change adds a helper class called InCallUiState, which keeps
track of some high-level state of the in-call UI that needs to persist
beyond a single pause/resume cycle of the InCallScreen.

This enables us to handle failed calls much more cleanly than before:
If there's a failure in the placeCall() sequence, the CallController
stashes away a failure code value in the "pending call status code"
field of the InCallUiState structure.  Then, when the InCallScreen comes
up, if that field contains an error code we do some UI-specific action
(like displaying a dialog) depending on the code.  See the "Error /
diagnostic indications" section of InCallUiState.java for the full
details.

Other details:

- Added a new class called "Constants", which currently contains the
  CallStatusCode enum, but which I'll eventually use for other app-wide
  constants too (for example the various EXTRA_* and intent action
  constants that are currently all over the place.)

- Removed references to Intent.ACTION_ANSWER (which has never actually
  been used for anything)

- Cleaned up confusing use of EXTRA_NEW_CALL_INTENT in
  OutgoingCallBroadcaster (bug 3325827)

- Finally bit the bullet and added an mApp field to InCallScreen (so we
  no longer need PhoneApp.getInstance() calls all over the place)

This change is just the first step of the InCallScreen/CallController
overhaul.  Basic calling works fine with this change, but some other
stuff still needs to be fixed:

  - OTASP-related behavior: (1) Handle the PERFORM_CDMA_PROVISIONING
    intent in the CallController, not the InCallScreen; (2) move the
    functions of InCallScreen.checkIsOtaCall() over to the CallController.

  - Implement the "emergency call from airplane mode" sequence inside the
    CallController (and get rid of the EmergencyCallHandler activity)

  - plus various other loose ends covered by TODO comments.

TESTED: "Basic telephony functionality" on Crespo:
  - outgoing calls
  - incoming calls
  - failed outgoing calls (airplane mode, no service, etc.)
  - bail out and return to InCallScreen while in-call
  - basic telephony states (two lines in use, hold/unhold, etc.)

Bug: 4194458
Bug: 3325827

Change-Id: I0523ef182a53fc7f252940d0b0e2804661dc9247